Change reset demo command to only refresh the database if the demo was accessed within the last hour. If the demo wasn't accessed within the last hour, there is no need to clean up

+++ app/Console/Commands/ResetDemo.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use App\Models\User;
+use Carbon\Carbon;
 use Illuminate\Console\Command;
 
 class ResetDemo extends Command
@@ -22,6 +23,15 @@ class ResetDemo extends Command
                 'email' => config('calendar.demo.email')
             ])
             ->first();
+
+        // Check last access
+        // If no login in last hour, no changes (except database)
+        // Prevents IDs from becoming unnecessarily large
+        if (Carbon::now()->gt(Carbon::parse($demoUser->last_login_at)->addHour())) {
+            $this->info('No sign-in since last reset');
+
+            return Command::SUCCESS;
+        }
 
         $demoUser->tags()->delete();
